requirements.txt
# core
flask
flask-sqlalchemy
flask-marshmallow
flask-smorest
flask-jwt-extended

# testing
pytest

# production
gunicorn
mariadb
